+++ src/meta/babasemeta/pyembed/binding_base.py
@@ -78,6 +78,6 @@ values = [
     _hooks.implicit_sign_out,  # kImplicitSignOutCall
     _hooks.login_adapter_get_sign_in_token_response,  # kLoginAdapterGetSignInTokenResponseCall
     _hooks.open_url_with_webbrowser_module,  # kOpenURLWithWebBrowserModuleCall
-    _env.on_native_module_import,  # kOnNativeModuleImportCall
+    _env.on_native_module_import,  # kEnvOnNativeModuleImportCall
     _env.on_main_thread_start_app,  # kOnMainThreadStartAppCall
 ]